Change DuplicateList to hashset
Duplicates are now stored in an hashset. Where duplicates are equal if
their VideoFiles are equal. Comparing file1 to file1 and file2 to
file2 but also file1 to file2 and file2 to file1. In case the files in
the duplicate are reversed.
