+++ VideoDedupConsole/VideoDedupConsole.csproj
@@ -48,8 +48,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="DuplicateWrapper.cs" />
-    <Compile Include="ThumbnailManager.cs" />
+    <Compile Include="DuplicateManager.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Properties\Settings.Designer.cs">
